requirements.txt
fastapi
uvicorn
whisper
ffmpeg-python
jinja2
python-multipart 